pyproject.toml
[tool.poetry]
name = "todolist-htmx-fastapi"
version = "0.1.0"
description = ""
authors = ["Guillaume Gauvrit <guillaume@gauvr.it>"]
readme = "README.md"
packages = [{include = "todolist", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
jinja2 = "^3.1.2"
hypercorn = "^0.14.3"
types-setuptools = "^67.7.0.2"
venusian = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
behave = "^1.2.6"
selenium = "^4.9.1"
black = "^23.3.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pyhamcrest = "^2.0.4"
httpx = "^0.24.0"
pytest-asyncio = "^0.21.0"

[[tool.mypy.overrides]]
module = "todolist.*"
disallow_any_generics = true
disallow_untyped_defs = true

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
